backend/messenger/requirements.txt
wheel
Pillow
django
djangorestframework
django-allauth
django-rest-auth
django-cors-headers
markdown
django-filter
psycopg2
pytz
channels_redis
